refactor data tables components
