+++ components/ui/data-table.tsx
@@ -1,14 +1,3 @@
-"use client"
-
-import {
-  ColumnDef,
-  flexRender,
-  getCoreRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
-} from "@tanstack/react-table"
-
 import {
   Table,
   TableBody,
@@ -21,36 +10,14 @@ import {
 import { cn } from "@/lib/utils"
 
 import { Button } from "./button"
+import { flexRender, Table as TableType } from "@tanstack/react-table"
 
-type Props<TData, TValue> = {
+type Props<TData> = {
   className?: string
-  data: TData[]
-  columns: ColumnDef<TData, TValue>[]
-  sorting?: { id: string; desc: boolean }[]
+  table: TableType<TData>
 }
 
-const DataTable = <TData, TValue>({
-  data,
-  columns,
-  className,
-  sorting = [],
-}: Props<TData, TValue>) => {
-  const table = useReactTable({
-    data,
-    columns,
-    getCoreRowModel: getCoreRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    initialState: {
-      pagination: {
-        pageSize: 15,
-      },
-    },
-    state: {
-      sorting,
-    },
-  })
-
+const DataTable = <TData,>({ className, table }: Props<TData>) => {
   return (
     <div className={cn("flex w-full flex-col gap-8", className)}>
       <Table>
@@ -90,7 +57,10 @@ const DataTable = <TData, TValue>({
             ))
           ) : (
             <TableRow>
-              <TableCell colSpan={columns.length} className="h-24 text-center">
+              <TableCell
+                colSpan={table.getHeaderGroups()[0].headers.length}
+                className="h-24 text-center"
+              >
                 No results.
               </TableCell>
             </TableRow>
